mrview/mrcolour: split off colourmap code from GUI code
to avoid dependence of mrcolour on Qt
