+++ src/gui/mrview/tool/connectome/connectome.cpp
@@ -24,7 +24,7 @@
 #include "algo/threaded_loop.h"
 #include "file/path.h"
 #include "gui/dialog/file.h"
-#include "gui/mrview/colourmap.h"
+#include "gui/mrview/colourmapper.h"
 #include "math/math.h"
 #include "math/rng.h"
