+++ src/gui/mrview/tool/tractography/track_scalar_file.cpp
@@ -16,7 +16,7 @@
 
 #include "gui/mrview/tool/tractography/track_scalar_file.h"
 #include "gui/dialog/file.h"
-#include "gui/mrview/colourmap.h"
+#include "gui/mrview/colourmapper.h"
 #include "gui/mrview/tool/tractography/tractogram.h"
 
 
@@ -53,7 +53,7 @@ namespace MR
           // Colourmap menu:
           colourmap_menu = new QMenu (tr ("Colourmap menu"), this);
 
-          ColourMap::create_menu (this, colourmap_group, colourmap_menu, colourmap_actions, false, false);
+          ColourMapper::create_menu (this, colourmap_group, colourmap_menu, colourmap_actions, false, false);
           connect (colourmap_group, SIGNAL (triggered (QAction*)), this, SLOT (select_colourmap_slot()));
           colourmap_actions[1]->setChecked (true);
