+++ src/gui/mrview/window.h
@@ -24,7 +24,7 @@
 #include "gui/gui.h"
 #include "gui/mrview/gui_image.h"
 #include "gui/opengl/font.h"
-#include "gui/mrview/colourmap.h"
+#include "gui/mrview/colourmapper.h"
 #include "gui/mrview/colourmap_button.h"
 
 
@@ -149,7 +149,7 @@ namespace MR
 
           GL::Area* glwidget () const { return glarea; }
           GL::Lighting& lighting () { return *lighting_; }
-          ColourMap::Renderer colourbar_renderer;
+          ColourMapper colourbar_renderer;
 
           static void add_commandline_options (MR::App::OptionList& options);
           static Window* main;
@@ -242,7 +242,7 @@ namespace MR
           Math::Versorf orient;
           float field_of_view;
           int anatomical_plane, annotations;
-          ColourMap::Position colourbar_position, tools_colourbar_position;
+          ColourMapper::Position colourbar_position, tools_colourbar_position;
           bool snap_to_image_axes_and_voxel;
 
           float background_colour[3];
@@ -298,7 +298,7 @@ namespace MR
                   *about_action,
                   *aboutQt_action;
 
-          static ColourMap::Position parse_colourmap_position_str (const std::string& position_str);
+          static ColourMapper::Position parse_colourmap_position_str (const std::string& position_str);
 
           void paintGL ();
           void initGL ();
